Update README.md
Donation is a noun and when stuck in front of the word platform it is a noun adjunct, which is a noun acting like an adjective. It makes less sense to use a verb in this role. People don't use this platform to make donates. Imagine if you'd built an 'assess platform', a 'meet tool' or purchased an 'inspect lamp'; in these cases the corresponding nouns assessment, meeting and inspection work better.